Move the slack "post user logged out" code to the domain layer.
