+++ slackhealthbot/core/exceptions.py
